Implements database querying through input fields.
Creates two new files in `src/` that define how the login page looks,
and a third in a new `servers/` folder that dictates how to query the
`user_info.db` database. Unlike previous front-end files,
`src/Loginpage.js` is class based, and is an extension of React's
`Component` class. See the file for more details.
Updates `README.md` to reflect the changes.