Implements both the user and game databases (#15)
* Added login page js file and css file.
Also added button to go to the login page.

* Configured index.js to link /login to the login.js file

* Added empty Userpage

* Updated index.js to match new userpage

* Initial commit for implementing user login.

Added `databases/` to store the user data, `servers/` to modularize
running the backend servers. Miscellaneous other files have been updated
to reflect the changes.

* Implemented dark mode theme and added search box

* Tried to add dark mode for real

* Implements database querying through input fields.

Creates two new files in `src/` that define how the login page looks,
and a third in a new `servers/` folder that dictates how to query the
`user_info.db` database. Unlike previous front-end files,
`src/Loginpage.js` is class based, and is an extension of React's
`Component` class. See the file for more details.
Updates `README.md` to reflect the changes.

* General code cleanup + renaming.

Rewrite all of the `src/` webpages to follow the same object-oriented
coding principles that `src/Loginpage.js` has, as well as convert most
variables to lowercase.

* Fixed invalid/unused headers

* Created images file for organization

* package-lock.json Updated

* Merge game database querying into authentication (#14)

* Add games database.

* Update `games.db` to follow proper formatting.

* Move database code out of `servers/`.

Only one `.js` file is needed to maintain all databases, so
`start_db.js` will suffice. Logic is now added such that the homepage
search bar actually queries the games database, and outputs to console.

TODO: Implement some sort of fuzzy matching for searching, as it
currently only accepts case-sensitive substrings as search queries.

* Update `README.md` to reflect most recent changes.

Also adds TODO items that will need to be implemented later.

* Fixes outdated command in `README.md`.

Updates the command for starting up the local database servers, now that
`servers/` has been deprecated.

* Gamepage (#16)

* Configured import statements in every file

* Deleted random files, started working on homepage

* Configured react-boostrap to gamepages

* Configured navigation bar

* Finished Gamepage outline

* Cleaned up code added comments for everything

* Made the Navbar its own function, slightly improved button spacing in a couple of places

* Fixed really silly error in the prior commit

* Fixed really silly error in the prior commit

* Fix remaining merge conflict errors.

Updates various files for bug fixes, rewrites code to be more
object-oriented, and updates some code blocks for more consistent style.
Front-end pages should all be extensions of the react `Component` class,
and should use the `render()` method to display their contents.

* Created basic version of the user page

Made a simple 3 column user page. Includes a section for the backlog, playthroughs, and friends list. Also includes the navbar at the top. Needs to be altered to use the database

* Pretty-ified buttons

Made the buttons on the home page properly spaced after that was lost in a merge
Switched to using an InputGroup for the navbar so the button integrates with the textbox, and changed search buttons to be consistently green.
This may have made some things off center but I can't tell, if you can please fix :)

Co-authored-by: Jonah Paten <jpaten@ucla.edu>
Co-authored-by: Jed Miguel <jedmig4@gmail.com>
Co-authored-by: Jed Miguel <54789413+jedjedm@users.noreply.github.com>