Implement most backend features (#26)
* Configured login page

* Configured sign up page

* Links `Signup.js` to `user_info.db`.

When a username/password combination is entered into the signup sheet, a
simple password check is enabled and a new JSON object is added to
`user_info.db`.

* Minor cosmetic changes before branching.

Rename a few variables, clean up code.

* Implement backend functions for followers.

Added two functions `/addfollower` and `/removefollower`, both of which
take a JSON with `user` and `follower` attributes. The functions
add/remove the `follower` from the given `user`.

* Implement backend function to set/change status.

The function takes in a JSON as input, which contains `user` and
`userStatus`, and sets the `user`'s status to `userStatus`.

* Add functions for adding/removing to/from backlog.

The corresponding APIs take in inputs of `user` and `game` as inputs in
a JSON, and the functions are `addtobacklog` and `removefrombacklog`.

* Add functions to add/remove to/from library.

These functions are nearly identical to the functions that manage the
backlog. Also fixes some incorrect comments.

Co-authored-by: Jed Miguel <jedmig4@gmail.com>